objtool: let find_reloc_by_dest find the in-order 1st one
The current code seems assuming there might be only one relocation to
the same offset.

For architetures like RISC-V, this is not true. The object file might
contain multipule relocations to the same offset, such as R_RISCV_ADD32
and R_RISCV_SUB32 pair, R_RISCV_CALL_PLT and R_RISCV_RELAX pair.

The in-order 1st one is mostly concerned which might refer to the
target instruction. The 2nd refer to a local symbol belong to the base
section itself, such as .discard, .ignore, .rodata sections. We can
ignore them.

Signed-off-by: Kai Zhang <zhangkai@iscas.ac.cn>